--- opensoundscape.ini
@@ -1,19 +1,0 @@
-[general]
-data_dir = /home/bmooreii/projects/opensoundscape/opensoundscape_data/sylatr
-train_file = train.csv
-predict_file = predict.csv
-db_name = sylatr_small
-db_sparse = True
-
-[spect_gen]
-binary_closing_kernel_height = 1
-binary_closing_kernel_width = 1
-binary_dilation_kernel_height = 1
-binary_dilation_kernel_width = 1
-median_filter_kernel_height = 1
-median_filter_kernel_width = 1
-small_objects_kernel_size = 25
-
-[model_fit]
-; template_pool = scripts/template_pool.csv
-labels_list = Sylatr_song
