+++ opensoundscape/model_fit/model_fit_algo/lasseck2013/model_fit_algo.py
@@ -13,23 +13,23 @@ from sklearn.metrics import confusion_matrix
 from sklearn.metrics import recall_score
 from sklearn.metrics import precision_score
 from sklearn.metrics import f1_score
-from modules.db_utils import init_client
-from modules.db_utils import close_client
-from modules.db_utils import cursor_item_to_data
-from modules.db_utils import cursor_item_to_stats
-from modules.db_utils import read_spectrogram
-from modules.db_utils import return_cursor
-from modules.db_utils import write_file_stats
-from modules.db_utils import write_model
-from modules.db_utils import get_model_fit_skip
-from modules.db_utils import set_model_fit_skip
-from modules.spect_gen import spect_gen
-from modules.view import extract_segments
-from modules.utils import return_cpu_count
-from modules.image_utils import apply_gaussian_filter
-from modules.image_utils import generate_raw_spectrogram
-from modules.utils import get_percent_from_section
-from modules.utils import get_template_matching_algorithm
+from opensoundscape.utils.db_utils import init_client
+from opensoundscape.utils.db_utils import close_client
+from opensoundscape.utils.db_utils import cursor_item_to_data
+from opensoundscape.utils.db_utils import cursor_item_to_stats
+from opensoundscape.utils.db_utils import read_spectrogram
+from opensoundscape.utils.db_utils import return_cursor
+from opensoundscape.utils.db_utils import write_file_stats
+from opensoundscape.utils.db_utils import write_model
+from opensoundscape.utils.db_utils import get_model_fit_skip
+from opensoundscape.utils.db_utils import set_model_fit_skip
+from opensoundscape.spect_gen.spect_gen import spect_gen
+from opensoundscape.view.view import extract_segments
+from opensoundscape.utils.image_utils import apply_gaussian_filter
+from opensoundscape.utils.image_utils import generate_raw_spectrogram
+from opensoundscape.utils.utils import return_cpu_count
+from opensoundscape.utils.utils import get_percent_from_section
+from opensoundscape.utils.utils import get_template_matching_algorithm
 from cv2 import matchTemplate as opencvMatchTemplate
 from cv2 import minMaxLoc
