+++ opensoundscape/predict/predict_algo/lasseck2013/predict_algo.py
@@ -1,16 +1,16 @@
 import pandas as pd
 import numpy as np
-from modules.db_utils import init_client
-from modules.db_utils import close_client
-from modules.db_utils import cursor_item_to_data
-from modules.db_utils import cursor_item_to_stats
-from modules.db_utils import read_spectrogram
-from modules.db_utils import recall_model
-from modules.db_utils import return_cursor
-from modules.db_utils import write_file_stats
-from modules.spect_gen import spect_gen
-from modules.view import extract_segments
-from modules.utils import return_cpu_count
+from opensoundscape.utils.db_utils import init_client
+from opensoundscape.utils.db_utils import close_client
+from opensoundscape.utils.db_utils import cursor_item_to_data
+from opensoundscape.utils.db_utils import cursor_item_to_stats
+from opensoundscape.utils.db_utils import read_spectrogram
+from opensoundscape.utils.db_utils import recall_model
+from opensoundscape.utils.db_utils import return_cursor
+from opensoundscape.utils.db_utils import write_file_stats
+from opensoundscape.spect_gen.spect_gen import spect_gen
+from opensoundscape.view.view import extract_segments
+from opensoundscape.utils.utils import return_cpu_count
 from scipy import stats
 from cv2 import matchTemplate, minMaxLoc
 from concurrent.futures import ProcessPoolExecutor
