+++ opensoundscape/scripts/generate-images.sh
