+++ opensoundscape/utils/image_utils.py
