+++ pyproject.toml
@@ -1,0 +1,50 @@
+[tool.poetry]
+name = "opensoundscape"
+version = "0.2.2"
+description = "Open source, scalable acoustic classification for ecology and conservation"
+authors = ["Justin Kitzes <justin.kitzes@pitt.edu>",
+           "Barry Moore <moore0557@gmail.com>",
+           "Tessa Rhinehart <tessa.rhinehart@gmail.com>"
+          ]
+license = "MIT"
+readme = "README.md"
+repository = "https://github.com/jkitzes/opensoundscape"
+
+[tool.poetry.dependencies]
+python = "^3.6"
+pandas = "=0.23.4"
+numpy = "=1.15.4"
+matplotlib = "=3.0.2"
+docopt = "=0.6.2"
+scipy = "=1.2.0"
+scikit-image = "=0.14.1"
+pymongo = "=3.7.2"
+librosa = "=0.6.2"
+numba = "=0.42.0" # Shouldn't need to define this explicitly, issue with librosa?
+opencv-python-headless = {version = "4.0.0.21",optional = true}
+
+[tool.poetry.dev-dependencies]
+pytest = "=4.2.0"
+pre-commit = "=1.14.2"
+black = "=18.9b0"
+
+[tool.poetry.scripts]
+opensoundscape = "opensoundscape.console:run"
+opso_dump_cross_corrs = "opensoundscape.scripts.dump_cross_correlations:run"
+opso_find_high_cross_corrs = "opensoundscape.scripts.find_high_cross_correlations:run"
+opso_find_important_templates = "opensoundscape.scripts.find_important_templates:run"
+
+[build-system]
+requires = ["poetry>=0.12"]
+build-backend = "poetry.masonry.api"
+
+[tool.black]
+line-length = 88
+py36 = true
+include = '\.pyi?$'
+exclude = '''
+/(
+    \.git
+  | dist
+)/
+'''
