--- requirements-dev.txt
@@ -1,11 +1,0 @@
-pandas==0.19.2
-numpy==1.14.2
-matplotlib==2.1.2
-docopt==0.6.2
-scipy==1.0.0
-scikit-image==0.13.1
-pymongo==3.4.0
-pytest==3.6.1
-librosa==0.6.2
-pre-commit==1.12.0
-black==18.9b0
