--- requirements-singularity.txt
@@ -1,2 +1,0 @@
-librosa==0.6.2
-scikit-image==0.13.1
